src/pyproject.toml
[tool.poetry]
name = "wallet-django"
version = "0.1.0"
description = ""
authors = ["Никита <modaravvv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
